+++ src/template/index.ts
@@ -1,3 +1,4 @@
 export * from "./builders";
 export * from "./ephemeral";
 export * from "./options";
+export * from "./status";