+++ src/template/options.ts
@@ -1,4 +1,4 @@
-import { Intents } from "discord.js";
+import { ClientOptions, CommandInteraction, Intents } from "discord.js";
 
 const f = Intents.FLAGS;
 
@@ -7,24 +7,41 @@ const f = Intents.FLAGS;
  *
  * GUILDS, GUILD_MESSAGES, GUILD_MESSAGE_REACTIONS, DIRECT_MESSAGES, DIRECT_MESSAGE_REACTIONS
  */
-export const CocoaIntents = {
-    intents: [
-        f.GUILDS,
-        f.GUILD_MESSAGES,
-        f.GUILD_MESSAGE_REACTIONS,
-        f.DIRECT_MESSAGES,
-        f.DIRECT_MESSAGE_REACTIONS,
-    ],
-    // To Accept DM Message
-    partials: ["CHANNEL"],
-};
+export const CocoaIntents: ClientOptions["intents"] = [
+    f.GUILDS,
+    f.GUILD_MESSAGES,
+    f.GUILD_MESSAGE_REACTIONS,
+    f.DIRECT_MESSAGES,
+    f.DIRECT_MESSAGE_REACTIONS,
+];
 
 /**
  * Template Intents with Voice, includes
  *
  * CocoaIntents + GUILD_VOICE_STATES
  */
-export const DJCocoaIntents = {
+export const DJCocoaIntents: ClientOptions["intents"] = [
     ...CocoaIntents,
-    intents: [...CocoaIntents.intents, f.GUILD_VOICE_STATES],
+    f.GUILD_VOICE_STATES,
+];
+
+/** Template ClientOptions used in Cocoa Grader*/
+export const CocoaOptions: ClientOptions = {
+    intents: CocoaIntents,
+    // To Accept DM
+    partials: ["MESSAGE"],
 };
+
+/** Template ClientOptions used in Harunon.js */
+export const DJCocoaOptions: ClientOptions = {
+    intents: DJCocoaIntents,
+    // To Accept DM
+    partials: ["MESSAGE"],
+};
+
+export function Author(ctx: CommandInteraction) {
+    return {
+        name: ctx.user.username,
+        iconURL: ctx.user.avatarURL() ?? "",
+    };
+}